Extract getUsername function as an exported helper function.
