+++ client/src/components/GameSelector/index.js
@@ -3,6 +3,7 @@ import { Card } from 'react-bootstrap';
 import { Link } from "react-router-dom";
 
 import { withFirebase } from '../Firebase';
+import { getUsername } from '../Helpers';
 
 import 'bootstrap/dist/css/bootstrap.css';
 import './GameSelector.css';
@@ -30,7 +31,7 @@ class GameSelector extends Component {
       let gameData = { ...gameDoc.data(), gameId: gameDoc.id, hasEnded: true };
 
       // get username of first player
-      gameData["startPlayer"] = await this.getUsername(gameData.players[0]);
+      gameData["startPlayer"] = await getUsername(gameData.players[0]);
 
       // check if current player is the host
       gameData["isHost"] = gameData.startPlayer === userDoc.data().username;
@@ -50,7 +51,7 @@ class GameSelector extends Component {
 
         // get username of current player if game is not over
         const currentPlayerIndex = gameData.currentPlayerIndex;
-        gameData["currentPlayer"] = await this.getUsername(gameData.players[currentPlayerIndex]);
+        gameData["currentPlayer"] = await getUsername(gameData.players[currentPlayerIndex]);
       }
 
       games.push(gameData);
@@ -58,24 +59,6 @@ class GameSelector extends Component {
     })
 
     this.setState({ loading: false });
-  }
-
-  getUsername = async uid => {
-    const options = {
-      method: 'POST',
-      headers: {
-        'Accept': 'application/json',
-        'Content-Type': 'application/json'
-      },
-      body: uid
-    };
-
-    const url = '/getUsername';
-
-    const response = await fetch(url, options);
-    const username = await response.text();
-
-    return username;
   }
 
   render() {
